LPS-90208 Move to remote package
